7/13 4

+++ controllers/videoControlloer.js
@@ -39,7 +39,6 @@ export const postUpload = async (req, res) => {
     });
     req.user.videos.push(newVideo.id);
     req.user.save();
-    console.log(newVideo);
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -47,7 +46,6 @@ export const videoDetail = async (req, res) => {
     const { params: { id } } = req;
     try {
         const video = await Video.findById(id).populate('creator');
-        console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.users);
@@ -58,7 +56,11 @@ export const getEditVideo = async (req, res) => {
     const { params: { id } } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+        }
     } catch (error) {
         res.redirect(routes.home);
     };
@@ -77,7 +79,12 @@ export const postEditVideo = async (req, res) => {
 export const deleteVideo = async (req, res) => {
     const { params: { id } } = req;
     try {
-        await Video.findOneAndRemove({ _id: id });
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({ _id: id });
+        }
     } catch (error) { };
     res.redirect(routes.home);
-}
+};
